+++ packages/create-sitecore-jss/src/templates/angular-xmcloud/src/app/components/promo/promo.component.ts
@@ -1,0 +1,8 @@
+import { Component } from '@angular/core';
+import { SxaComponent } from '../sxa.component';
+
+@Component({
+  selector: 'app-promo',
+  templateUrl: './promo.component.html',
+})
+export class PromoComponent extends SxaComponent {}