+++ packages/sitecore-jss-angular/src/components/link.directive.ts
@@ -76,11 +76,15 @@ export class LinkDirective implements OnChanges {
     }
   }
 
+  private shouldRender() {
+    return this.field && (this.field.href || this.field.value?.href || this.field.text);
+  }
+
   private updateView() {
     const field = this.field;
     if (this.editable && field && field.editableFirstPart && field.editableLastPart) {
       this.renderInlineWrapper(field.editableFirstPart, field.editableLastPart);
-    } else if (field && (field.href || field.value)) {
+    } else if (this.shouldRender()) {
       const props = field.href ? field : field.value;
 
       const linkText = field.text || field.value?.text || field.href || field.value?.href;